requirements.txt
flask
flask-cors
google-generativeai
faiss-cpu
numpy
pymupdf
python-docx
gunicorn
flask_sqlalchemy
flask-jwt-extended
google-genai